requirements.txt
accelerate==0.23.0
alembic==1.12.0
beautifulsoup4==4.12.2
botorch==0.9.2
Bottleneck @ file:///work/ci_py311/bottleneck_1676827011612/work
brotlipy==0.7.0
certifi==2023.7.22
cffi @ file:///work/ci_py311/cffi_1676822533496/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click @ file:///home/conda/feedstock_root/build_artifacts/click_1692311806742/work
cloudpickle @ file:///home/conda/feedstock_root/build_artifacts/cloudpickle_1697464713350/work
colorlog==6.7.0
contourpy==1.1.1
cryptography @ file:///croot/cryptography_1694444244250/work
cycler==0.12.1
cytoolz @ file:///work/ci_py311/cytoolz_1676830405847/work
dask @ file:///home/conda/feedstock_root/build_artifacts/dask-core_1697245951516/work
dataclasses @ file:///home/conda/feedstock_root/build_artifacts/dataclasses_1628958434797/work
datasets==1.11.0
detectors==0.1.10
dill @ file:///home/conda/feedstock_root/build_artifacts/dill_1690101045195/work
einops @ file:///home/conda/feedstock_root/build_artifacts/einops_1696154234947/work
faiss-cpu==1.7.4
filelock @ file:///work/ci_py311/filelock_1676823024085/work
fonttools==4.43.1
fsspec @ file:///home/conda/feedstock_root/build_artifacts/fsspec_1695417469930/work
gdown==4.7.1
gmpy2 @ file:///work/ci_py311/gmpy2_1676839849213/work
gpytorch==1.11
greenlet==3.0.0
huggingface-hub==0.18.0
idna @ file:///work/ci_py311/idna_1676822698822/work
imagecodecs @ file:///work/ci_py311_2/imagecodecs_1679336294425/work
imageio @ file:///home/conda/feedstock_root/build_artifacts/imageio_1696854106455/work
importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1688754491823/work
jaxtyping==0.2.23
Jinja2 @ file:///work/ci_py311/jinja2_1676823587943/work
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1691577114857/work
kiwisolver==1.4.5
lazy_loader @ file:///home/conda/feedstock_root/build_artifacts/lazy_loader_1692295373316/work
linear-operator==0.5.1
littleutils==0.2.2
locket @ file:///home/conda/feedstock_root/build_artifacts/locket_1650660393415/work
Mako==1.2.4
MarkupSafe @ file:///work/ci_py311/markupsafe_1676823507015/work
matplotlib==3.8.0
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mpmath @ file:///croot/mpmath_1690848262763/work
multipledispatch==1.0.0
multiprocess @ file:///croot/multiprocess_1692294385131/work
networkx @ file:///croot/networkx_1690561992265/work
numexpr @ file:///croot/numexpr_1696515281613/work
numpy @ file:///croot/numpy_and_numpy_base_1695830428084/work/dist/numpy-1.26.0-cp311-cp311-linux_x86_64.whl#sha256=184178a893ff0b5e0cb1b0b7a0dddff49dc91424c4e4a01e29bcb05f27a698a6
ogb==1.3.6
opt-einsum==3.3.0
optuna==3.4.0
outdated==0.2.2
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1696202382185/work
pandas @ file:///croot/pandas_1697477158611/work/dist/pandas-2.1.1-cp311-cp311-linux_x86_64.whl#sha256=ed5518777956b76c6290791e98ec9ff6ea64a621c757052a7697aa4252bc0d9b
partd @ file:///home/conda/feedstock_root/build_artifacts/partd_1695667515973/work
Pillow @ file:///croot/pillow_1696580024257/work
psutil==5.9.6
pyarrow==11.0.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyOpenSSL @ file:///croot/pyopenssl_1690223430423/work
pyparsing==3.1.1
pyro-api==0.1.2
pyro-ppl==1.8.6
PySocks @ file:///work/ci_py311/pysocks_1676822712504/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1693930252784/work
PyWavelets @ file:///work/ci_py311/pywavelets_1676838693522/work
PyYAML @ file:///work/ci_py311/pyyaml_1676827445582/work
regex @ file:///croot/regex_1696515298636/work
requests @ file:///croot/requests_1690400202158/work
sacremoses==0.0.43
safetensors @ file:///croot/safetensors_1697478848712/work
scikit-image @ file:///croot/scikit-image_1682528309703/work
scikit-learn==1.3.1
scipy @ file:///croot/scipy_1696543286448/work/dist/scipy-1.11.3-cp311-cp311-linux_x86_64.whl#sha256=f14db5b9e7e3d9de204ef7756f1a1f40f0983293019ed8eacb32191e2f658f9f
seaborn==0.13.0
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
soupsieve==2.5
SQLAlchemy==2.0.22
sympy @ file:///work/ci_py311_2/sympy_1679339311852/work
threadpoolctl==3.2.0
tifffile @ file:///home/conda/feedstock_root/build_artifacts/tifffile_1617956128193/work
timm==0.8.19.dev0
tokenizers @ file:///croot/tokenizers_1696539574676/work
toolz @ file:///home/conda/feedstock_root/build_artifacts/toolz_1657485559105/work
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
tqdm==4.66.1
transformers @ file:///home/conda/feedstock_root/build_artifacts/transformers_1689730109002/work
triton==2.1.0
typeguard==2.13.3
typing_extensions @ file:///croot/typing_extensions_1690297465030/work
tzdata @ file:///home/conda/feedstock_root/build_artifacts/python-tzdata_1680081134351/work
urllib3 @ file:///croot/urllib3_1686163155763/work
wilds==2.0.0
xxhash @ file:///work/ci_py311/python-xxhash_1676842384694/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1695255097490/work
